Add secretkey verify

+++ src/trainings/trainings.controller.ts
@@ -2,12 +2,16 @@ import {
   Controller,
   Post,
   Body,
+  Headers,
   Get,
   Param,
   Patch,
   Delete,
 } from '@nestjs/common';
 import { TrainingsService } from './trainings.service';
+
+// tslint:disable-next-line: no-var-requires
+require('dotenv').config();
 
 @Controller('trainings')
 export class TrainingsController {
@@ -23,110 +27,161 @@ export class TrainingsController {
     @Body('time') time: number | null,
     @Body('userId') userId: string,
     @Body('type') type: string,
+    @Headers() headers,
   ) {
-    const generatedId = await this.trainingsService.insertTraining({
-      trainingDate,
-      colorTag,
-      description,
-      distance,
-      calories,
-      time,
-      userId,
-      type,
-    });
-    return { id: generatedId };
+    if (headers.key === process.env.SECRET_KEY) {
+      const generatedId = await this.trainingsService.insertTraining({
+        trainingDate,
+        colorTag,
+        description,
+        distance,
+        calories,
+        time,
+        userId,
+        type,
+      });
+      return { id: generatedId };
+    } else {
+      return null;
+    }
   }
 
   @Get()
-  async getAllTranings() {
-    const allTrainings = await this.trainingsService.getTranings();
-    return allTrainings;
+  async getAllTranings(@Headers() headers) {
+    if (headers.key === process.env.SECRET_KEY) {
+      const allTrainings = await this.trainingsService.getTranings();
+      return allTrainings;
+    } else {
+      return null;
+    }
   }
 
   @Get('/user/:userId')
-  async getTranings(@Param('userId') userId: string) {
-    const userTranings = await this.trainingsService.getTrainingsForUser({
-      userId,
-    });
-    return userTranings;
+  async getTranings(@Param('userId') userId: string, @Headers() headers) {
+    if (headers.key === process.env.SECRET_KEY) {
+      const userTranings = await this.trainingsService.getTrainingsForUser({
+        userId,
+      });
+      return userTranings;
+    } else {
+      return null;
+    }
   }
 
   @Get('/user/:userId/id/:id')
-  async getTraning(@Param('id') id: string, @Param('userId') userId: string) {
-    const training = await this.trainingsService.getSingleTraining({
-      userId,
-      id,
-    });
-    return training;
+  async getTraning(
+    @Param('id') id: string,
+    @Param('userId') userId: string,
+    @Headers() headers,
+  ) {
+    if (headers.key === process.env.SECRET_KEY) {
+      const training = await this.trainingsService.getSingleTraining({
+        userId,
+        id,
+      });
+      return training;
+    } else {
+      return null;
+    }
   }
 
   @Get('/first/user/:userId')
   async getFirstTraning(
     @Param('id') traningId: string,
     @Param('userId') userId: string,
+    @Headers() headers,
   ) {
-    const training = await this.trainingsService.getFirstTrainingForUser({
-      userId,
-    });
-    return training;
+    if (headers.key === process.env.SECRET_KEY) {
+      const training = await this.trainingsService.getFirstTrainingForUser({
+        userId,
+      });
+      return training;
+    } else {
+      return null;
+    }
   }
 
   @Get('/last/user/:userId')
   async getLastTraning(
     @Param('id') traningId: string,
     @Param('userId') userId: string,
+    @Headers() headers,
   ) {
-    const training = await this.trainingsService.getLastTrainingForUser({
-      userId,
-    });
-    return training;
+    if (headers.key === process.env.SECRET_KEY) {
+      const training = await this.trainingsService.getLastTrainingForUser({
+        userId,
+      });
+      return training;
+    } else {
+      return null;
+    }
   }
 
   @Get('/calories/user/:userId/year/:year')
   async getTheLargestAmountOfCalories(
     @Param('userId') userId: string,
     @Param('year') year: string,
+    @Headers() headers,
   ) {
-    const training = await this.trainingsService.getTheLargestAmountOfCaloriesForUser(
-      year,
-      { userId },
-    );
-    return training;
+    if (headers.key === process.env.SECRET_KEY) {
+      const training = await this.trainingsService.getTheLargestAmountOfCaloriesForUser(
+        year,
+        { userId },
+      );
+      return training;
+    } else {
+      return null;
+    }
   }
 
   @Get('/distance/user/:userId/year/:year')
   async getTheLargestAmountOfDistance(
     @Param('userId') userId: string,
     @Param('year') year: string,
+    @Headers() headers,
   ) {
-    const training = await this.trainingsService.getTheLargestAmountOfDistanceForUser(
-      year,
-      { userId },
-    );
-    return training;
+    if (headers.key === process.env.SECRET_KEY) {
+      const training = await this.trainingsService.getTheLargestAmountOfDistanceForUser(
+        year,
+        { userId },
+      );
+      return training;
+    } else {
+      return null;
+    }
   }
 
   @Get('/time/user/:userId/year/:year')
   async getTheLargestAmountOfTime(
     @Param('userId') userId: string,
     @Param('year') year: string,
+    @Headers() headers,
   ) {
-    const training = await this.trainingsService.getTheLargestAmountOfTimeForUser(
-      year,
-      { userId },
-    );
-    return training;
+    if (headers.key === process.env.SECRET_KEY) {
+      const training = await this.trainingsService.getTheLargestAmountOfTimeForUser(
+        year,
+        { userId },
+      );
+      return training;
+    } else {
+      return null;
+    }
   }
 
   @Get('/sum/user/:userId/year/:year')
   async getSumTraningDataByYear(
     @Param('userId') userId: string,
     @Param('year') year: string,
+    @Headers() headers,
   ) {
-    const result = await this.trainingsService.sumTraingsDataByYear(year, {
-      userId,
-    });
-    return result;
+    if (headers.key === process.env.SECRET_KEY) {
+      const result = await this.trainingsService.sumTraingsDataByYear(year, {
+        userId,
+      });
+      return result;
+    } else {
+      return null;
+    }
   }
 
   @Get('/sum/user/:userId/year/:year/month/:month')
@@ -134,13 +189,18 @@ export class TrainingsController {
     @Param('userId') userId: string,
     @Param('year') year: string,
     @Param('month') month: string,
+    @Headers() headers,
   ) {
-    const result = await this.trainingsService.sumTraingsDataByMonth(
-      year,
-      month,
-      { userId },
-    );
-    return result;
+    if (headers.key === process.env.SECRET_KEY) {
+      const result = await this.trainingsService.sumTraingsDataByMonth(
+        year,
+        month,
+        { userId },
+      );
+      return result;
+    } else {
+      return null;
+    }
   }
 
   @Patch('/user/:userId/id/:id')
@@ -153,26 +213,36 @@ export class TrainingsController {
     @Body('calories') calories: number,
     @Body('time') time: number,
     @Body('type') type: string,
+    @Headers() headers,
   ) {
-    await this.trainingsService.updateTraining({
-      userId,
-      id,
-      colorTag,
-      description,
-      distance,
-      calories,
-      time,
-      type,
-    });
-    return null;
+    if (headers.key === process.env.SECRET_KEY) {
+      await this.trainingsService.updateTraining({
+        userId,
+        id,
+        colorTag,
+        description,
+        distance,
+        calories,
+        time,
+        type,
+      });
+      return null;
+    } else {
+      return null;
+    }
   }
 
   @Delete('/user/:userId/id/:id')
   async removeTraining(
     @Param('userId') userId: string,
     @Param('id') id: string,
+    @Headers() headers,
   ) {
-    await this.trainingsService.deleteTraining({ userId, id });
-    return null;
+    if (headers.key === process.env.SECRET_KEY) {
+      await this.trainingsService.deleteTraining({ userId, id });
+      return null;
+    } else {
+      return null;
+    }
   }
 }
